feat: 공지사항 상세 파일 다운로드 함수 분리 및 수정 (#8)
