+++ app/(landing)/notices/_ui/notice-detail/index.tsx
@@ -3,36 +3,17 @@
 import { DownloadIcon } from '@/public/icons'
 import classNames from 'classnames/bind'
 
-import axios from '@/shared/api/axios'
-
+import { handleNoticeFileDownload } from '../../_api/get-notice-file'
 import useNoticeDetail from '../../_hooks/use-notice-detail'
 import styles from './styles.module.scss'
 
 const cx = classNames.bind(styles)
 
 const NoticeDetail = ({ noticeId }: { noticeId: number }) => {
-  const { data: notice, isLoading } = useNoticeDetail(noticeId)
+  const { data: notice } = useNoticeDetail(noticeId)
 
-  const handleSave = async (fileName: string, noticeFileId: number) => {
-    try {
-      const response = await axios.get(`/api/files/download/${noticeFileId}`, {
-        responseType: 'blob',
-      })
-
-      const blob = new Blob([response.data])
-      const url = window.URL.createObjectURL(blob)
-
-      const link = document.createElement('a')
-      link.href = url
-      link.download = fileName
-      document.body.appendChild(link)
-      link.click()
-
-      link.remove()
-      window.URL.revokeObjectURL(url)
-    } catch (err) {
-      console.error('파일 다운로드 중 오류 발생:', err)
-    }
+  const handleSave = (fileName: string, noticeFileId: number) => {
+    handleNoticeFileDownload(fileName, noticeId, noticeFileId)
   }
 
   return (
@@ -58,7 +39,10 @@ const NoticeDetail = ({ noticeId }: { noticeId: number }) => {
               >
                 {file.fileName}
               </div>
-              <DownloadIcon className={cx('file-download')} onClick={() => handleSave} />
+              <DownloadIcon
+                className={cx('file-download')}
+                onClick={() => handleSave(file.fileName, file.noticeFileId)}
+              />
             </div>
           ))}
         </div>
